refactor: update file and folder names
