+++ components/compositions/booking-modal/payment-info.tsx
